backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
httpx==0.28.1
feedparser==6.0.11
apscheduler==3.10.4
python-dotenv==1.0.1
yfinance
anthropic
